// apps/Backend/src/types.ts
export type TimeDurationCandel = "1m" | "5m" | "15m" | "1h" | "1d" | "1w";
// ...
export interface Candle {
  time: number;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: string;
}
export type ValidSymbol = "BTC" | "ETH" | "SOL";